Explicited nb of imgs for OF optimization with Magdeburg dataset and fixed bug in iterative 2D OF due to type inconsistency

+++ Image_prediction/1. Optical flow calculation/iterative_2Dof.m
@@ -10,7 +10,7 @@ function [ u_temp ] = iterative_2Dof( I, J, u_temp, OF_par )
     grad_I = spatial_grad_2D(I, OF_par);
     for k = 1:OF_par.nb_iter
         fprintf('\t\t refinement iteration n°%d \n', k); 
-        u_temp = u_temp + OF_LK_2D( grad_I, translate2DIm(J, u_temp) - I, OF_par);
+        u_temp = u_temp + OF_LK_2D( grad_I, translate2DIm(J, u_temp) - double(I), OF_par);
     end
 
 end
